+++ docs/conf.py
@@ -36,7 +36,7 @@ copyright = '2024, Paul Bourne'
 author = 'Paul Bourne'
 
 # The short X.Y version
-version = '0.7.4'
+version = '0.8.0'
 # The full version, including alpha/beta/rc tags
 release = version