+++ pycatia/drafting_interfaces/drawing_arrow.py
@@ -140,6 +140,37 @@ class DrawingArrow(AnyObject):
         return self.drawing_arrow.NbPoint
 
     @property
+    def scale_on_extremities(self) -> bool:
+        """
+        Introduced in V5-6R2018.
+
+        .. note::
+            :class: toggle
+
+            CAA V5 Visual Basic Help (2024-08-20 15:47:12.690381)
+                | Property ScaleOnExtremities() As boolean
+                |     Returns or sets the scale on extremities mode.
+                |
+                |     Example:
+                |         This example retrieves the target element of head side on MyArrow
+                |         drawing arrow.
+                |
+                |          oScaleOnExtremities = MyArrow.ScaleOnExtremities
+
+        :rtype: bool
+        """
+
+        return self.drawing_arrow.ScaleOnExtremities
+
+    @scale_on_extremities.setter
+    def scale_on_extremities(self, value: bool):
+        """
+        :param bool value:
+        """
+
+        self.drawing_arrow.ScaleOnExtremities = value
+
+    @property
     def tail_symbol(self) -> int:
         """
         .. note::