+++ pycatia/part_interfaces/hole.py
@@ -167,6 +167,44 @@ class Hole(SketchBasedShape):
         self.hole.BottomType = value
 
     @property
+    def counter_drilled_mode(self) -> int:
+        """
+
+        Introduced in V5-6R2018.
+
+        .. note::
+            :class: toggle
+
+            CAA V5 Visual Basic Help (2024-08-20 16:04:57.203445)
+                | Property CounterDrilledMode() As CatCDHoleMode
+                |     Returns the mode of a counterdrilled hole .
+                |
+                |     Returns:
+                |         CDMode Value of the counterdrilled mode (see CatCDHoleMode for list of
+                |         possible types)
+                |
+                |         Example:
+                |             The following example returns in CDMode the CDMode of hole
+                |             firsthole:
+                |
+                |              Set CDMode = firsthole.CounterDrilledMode
+
+        See enumeration type: cat_cd_hole_mode
+
+        :rtype: int
+        """
+
+        return self.hole.CounterDrilledMode
+
+    @counter_drilled_mode.setter
+    def counter_drilled_mode(self, value: int):
+        """
+        :param int value:
+        """
+
+        self.hole.CounterDrilledMode = value
+
+    @property
     def counter_sunk_mode(self) -> int:
         """
         .. note::